feat: setting alias should print progressbar
